+++ lib/qc/modules/qc_tester/calculator_media_capture/providers/calculator_media_capture_provider.dart
@@ -14,6 +14,11 @@ class CalculatorMediaCaptureProvider extends CalculatorServiceInitProvider {
   DeviceMediaResponse? deviceMediaResponse;
   bool isDataLoading = true;
   String? errorMessage;
+  final bool? isComingFromCalJourney;
+  final String deviceBarcode;
+  final int? categoryId;
+
+  CalculatorMediaCaptureProvider(this.deviceBarcode, this.isComingFromCalJourney, {this.categoryId}) : super();
 
   bool isAllMediaUpLoaded() {
     if (Validator.isListNullOrEmpty(deviceMediaResponse?.imageList)) {
@@ -32,11 +37,6 @@ class CalculatorMediaCaptureProvider extends CalculatorServiceInitProvider {
     return Provider.of<CalculatorMediaCaptureProvider>(context, listen: listen);
   }
 
-  @override
-  void onServiceInitialized() {
-    _getDeviceMedia();
-  }
-
   void saveMediaList() {
     CalculatorDataHolderModel().mediaList = deviceMediaResponse?.imageList
         ?.map((e) => MediaSubmitRequest(
@@ -44,29 +44,26 @@ class CalculatorMediaCaptureProvider extends CalculatorServiceInitProvider {
         .toList();
   }
 
-  _getDeviceMedia() {
-    service.getDeviceMedia(CalculatorDataHolderModel().deviceBarcode ?? "").listen((event) {
-      if (event != null) {
+  void getDeviceMedia({VoidCallback? onMoveToNextScreen}) {
+    service.getDeviceMedia(deviceBarcode, categoryId: categoryId).listen((event) {
+      isDataLoading = false;
+      if (Validator.isListNullOrEmpty(event?.imageList) && Validator.isTrue(isComingFromCalJourney)) {
+        onMoveToNextScreen?.call();
+      } else {
         deviceMediaResponse = event;
+        notifyListeners();
       }
     }, onError: (error) {
       String msg = ApiErrorHelper.getErrorMessage(error) ?? "Something went wrong";
-      Logger.debug('mydebug------CalculatorMediaCaptureProvider._getDeviceMedia', [msg]);
       errorMessage = msg;
-    }, onDone: () {
       isDataLoading = false;
       notifyListeners();
     });
   }
 
-  bool isCaptureMediaJourney() {
-    return CalculatorDataHolderModel().isCaptureDeviceMediaJourney;
-  }
-
   Future<bool> submitDeviceMedia() {
     var completer = Completer<bool>();
-    service.submitDeviceMedia(CalculatorDataHolderModel().mediaList, CalculatorDataHolderModel().deviceBarcode).listen(
-        (event) {
+    service.submitDeviceMedia(CalculatorDataHolderModel().mediaList, deviceBarcode).listen((event) {
       if (event != null) {
         completer.complete(true);
       } else {