+++ lib/src/libraries/firebase/remote_config_helper.dart
@@ -5,8 +5,6 @@ import 'package:firebase_remote_config/firebase_remote_config.dart';
 import 'package:flutter_trc/src/utils/connectivity_util.dart';
 
 class AppRemoteConfig {
-  static const KEY_IS_CAPTURE_MEDIA_MANDATORY_IN_QC = "key_is_capture_media_mandatory_in_qc";
-  static const KEY_IS_CAPTURE_MEDIA_MANDATORY_IN_TRC = "key_is_capture_media_mandatory_in_trc";
   static const KEY_IS_FORCE_SERVER_RULE_EXECUTOR = "key_is_force_server_rule_executor";
   static const KEY_IS_ENABLE_RULE_EXE_TEST_MODE = "key_is_enable_rule_exe_test_mode";
   static const KEY_VIDEO_RECORD_DURATION_IN_SEC = "key_video_record_duration_in_sec";
@@ -17,8 +15,6 @@ class AppRemoteConfig {
   static const KEY_IMEI_READER_TIMEOUT_SEC = "key_imei_reader_timeout_sec";
 
   static const Map<String, dynamic> DEFAULT_CONFIG = {
-    KEY_IS_CAPTURE_MEDIA_MANDATORY_IN_QC: true,
-    KEY_IS_CAPTURE_MEDIA_MANDATORY_IN_TRC: false,
     KEY_IS_FORCE_SERVER_RULE_EXECUTOR: true,
     KEY_IS_ENABLE_RULE_EXE_TEST_MODE: true,
     KEY_VIDEO_RECORD_DURATION_IN_SEC: 1200,