+++ lib/src/utils/media_upload/providers/image_upload_provider.dart
@@ -1,28 +1,26 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:core_widgets/core_widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_trc/src/utils/connectivity_util.dart';
-import 'package:flutter_trc/src/utils/media_upload/models/image_upload_service_type_enum.dart';
+import 'package:flutter_trc/src/utils/media_upload/providers/media_upload_service_init_provider.dart';
 import 'package:path/path.dart' as path;
 import 'package:provider/provider.dart';
 
 import '../media_optimiser_utils.dart';
 
-class ImageUploadProvider extends CshChangeNotifier {
+class ImageUploadProvider extends MediaUploadServiceInitProvider {
   bool isDataLoading = false;
   String? s3Url = "";
-  ImageUploadServiceType serviceType;
 
-  ImageUploadProvider({this.serviceType = ImageUploadServiceType.image_optimize, this.s3Url});
+  ImageUploadProvider({this.s3Url});
 
   static ImageUploadProvider of(BuildContext context, {bool listen = true}) {
     return Provider.of<ImageUploadProvider>(context, listen: listen);
   }
 
   Future<String> uploadImage(File file) async {
-    var isConnected= await ConnectivityUtil.checkConnectivity();
+    var isConnected = await ConnectivityUtil.checkConnectivity();
     if (isConnected == false) {
       return Future.error("No Internet Connection");
     }
@@ -31,8 +29,7 @@ class ImageUploadProvider extends CshChangeNotifier {
     notifyListeners();
     var completer = Completer<String>();
     String fileName = path.basename(file.path);
-    MediaUploadUtil(service: serviceType.service).uploadMediaWithType(mediaFile: file, fileName: fileName).then(
-        (value) {
+    MediaUploadUtil(service: mediaUploadService).uploadMediaWithType(mediaFile: file, fileName: fileName).then((value) {
       if (value.isNotEmpty) {
         s3Url = value;
         completer.complete(s3Url);