+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # TRC Releases
 
+# v4.0.3+69 (XX-02-2025)
+
+### Feature:
+
+* `QC/TRC - Move Image Capture under Manual Testing`[(SCM-8886)](https://reglobe.atlassian.net/browse/SCM-8886)
+* `QC - Device Details - changes in design and add stock movement`[(SCM-8882)](https://reglobe.atlassian.net/browse/SCM-8882)
+
+
 # v4.0.2+68 (12-02-2025)
 
 ### Feature: